Simplifying our customization files
Our customizations are not extensive or involved. I had set
up this environment to be used for a considerably more involved
process that never materialized. If an updated, more involved
environment is needed in the future, at this time I would suggest
using ES modules and webpack with babel. As of now this little bit
of JS and CSS does not need Grunt and SASS and all that.